fix python run error
