+++ python/runtime/optimize/__init__.py
@@ -11,6 +11,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from runtime.optimize.optflow_submit import run_optimize_on_optflow
 from runtime.optimize.optimize import (generate_model_with_data_frame,
-                                       run_optimize_locally,
-                                       run_optimize_on_optflow)
+                                       run_optimize_locally)
